rrr: convert rpc/completion_tracker to module (rrr.completion_tracker)
completion_tracker.cpp module; .hpp shim. Exports CompletionTracker
class (Mutex<HashMap<...>> backing) that records completed XIDs and
their cached results for after-reconnect status queries.

Imports rrr.idempotency for CachedResponse type.

33 modules. 77.98 s wallclock = 4.62× baseline. Approaching the 5×
ceiling (~84 s) — slowing down per-module growth tracks the
prediction that wide-fan-out modules (everything via client.hpp)
extend the slowest TU.
